requirements.txt
django  # High-level Python web framework that encourages rapid development and clean, pragmatic design.
djangorestframework  # Powerful and flexible toolkit for building Web APIs in Django.
requests  # Simple, yet elegant HTTP library for Python.
django-cors-headers  # Django app for handling the server headers required for Cross-Origin Resource Sharing (CORS).
python-dotenv  # Reads key-value pairs from a .env file and can set them as environment variables.
psycopg2-binary  # PostgreSQL database adapter for Python.
gunicorn
azure-monitor-opentelemetry>=1.2.0  # OpenTelemetry instrumentation for Azure Monitor.
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-django>=0.42b0
opencensus-ext-azure
azure-ai-projects